requirements.txt
Flask>=2.3,<4
Flask-SQLAlchemy>=3.0,<4
Flask-Login>=0.6,<1
Flask-WTF>=1.1,<2
pyspark>=3.4,<4
pandas>=2.0,<3
plotly>=5.15,<6
numpy>=1.24,<3
pyodbc>=4.0,<6
Werkzeug>=2.3,<4
email-validator>=2.0,<3
Flask-Migrate>=4.0,<5
setuptools
python-dotenv>=1.0
